Added vite to the project for faster development server
changed the index.html to root directory

changed js files to jsx
