+++ package.json
@@ -1,51 +1,42 @@
 {
-    "name": "react-portfolio",
-    "version": "0.1.0",
-    "private": true,
-    "dependencies": {
-        "@chakra-ui/icons": "^2.1.1",
-        "@chakra-ui/react": "^2.8.2",
-        "@emotion/react": "^11.11.3",
-        "@emotion/styled": "^11.11.0",
-        "bootstrap": "^5.2.3",
-        "emailjs-com": "^3.2.0",
-        "framer-motion": "^10.18.0",
-        "react": "^18.2.0",
-        "react-bootstrap": "^2.7.0",
-        "react-dom": "^18.2.0",
-        "react-helmet-async": "^1.0.7",
-        "react-icons": "^4.1.0",
-        "react-router-dom": "^6.6.2",
-        "react-scripts": "^5.0.1",
-        "react-transition-group": "^4.4.2",
-        "styled-components": "^6.1.8",
-        "typewriter-effect": "^2.18.2",
-        "web-vitals": "^3.1.1"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
-        "predeploy": "yarn build && cp build/index.html build/404.html",
-        "deploy": "gh-pages -d build"
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    }
+  "name": "folio",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "@chakra-ui/icons": "^2.1.1",
+    "@chakra-ui/react": "^2.8.2",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "bootstrap": "^5.2.3",
+    "emailjs-com": "^3.2.0",
+    "framer-motion": "^10.18.0",
+    "react-bootstrap": "^2.7.0",
+    "react-helmet-async": "^1.0.7",
+    "react-icons": "^4.1.0",
+    "react-router-dom": "^6.6.2",
+    "react-scripts": "^5.0.1",
+    "react-transition-group": "^4.4.2",
+    "styled-components": "^6.1.8",
+    "typewriter-effect": "^2.18.2",
+    "web-vitals": "^3.1.1"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "vite": "^5.1.4"
+  }
 }
