--- src/app/routes.js
@@ -1,54 +1,0 @@
-import React from "react";
-import { Route, Routes, Link } from "react-router-dom";
-import withRouter from "../hooks/withRouter";
-import { Home } from "../pages/home";
-import { Portfolio } from "../pages/portfolio";
-import { ContactUs } from "../pages/contact";
-import { About } from "../pages/about";
-import { Socialicons } from "../components/socialicons";
-import { CSSTransition, TransitionGroup } from "react-transition-group";
-import Project from "../pages/Project/Project"
-import dataportfolio from "../data/dataportfolio"; // Import your project data
-
-const AnimatedRoutes = withRouter(({ location }) => (
-  <TransitionGroup>
-    <CSSTransition
-      key={location.key}
-      timeout={{
-        enter: 400,
-        exit: 400,
-      }}
-      classNames="page"
-      unmountOnExit
-    >
-      <Routes location={location}>
-        <Route path="/" element={<Home />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/portfolio" element={<Portfolio />} />
-        <Route path="/projects" element={<Portfolio />} /> 
-        
-        {dataportfolio.map((project) => (
-          <Route
-            key={project.id}
-            path={`/projects/${project.id}`}
-            element={<Project project={project} />}
-          />
-        ))}
-
-        <Route path="/contact" element={<ContactUs />} />
-        <Route path="*" element={<Home />} />
-      </Routes>
-    </CSSTransition>
-  </TransitionGroup>
-));
-
-function AppRoutes() {
-  return (
-    <div className="s_c">
-      <AnimatedRoutes />
-      <Socialicons />
-    </div>
-  );
-}
-
-export default AppRoutes;
