+++ src/components/socialicons/index.jsx
@@ -1,0 +1,66 @@
+
+import "./style.css";
+import {
+  FaGithub,
+  FaTwitter,
+  FaFacebookF,
+  FaLinkedin,
+  FaYoutube,
+  FaTwitch,
+} from "react-icons/fa";
+import { socialprofils } from "../../content_option";
+import { Container } from "react-bootstrap"
+
+export const Socialicons = () => {
+  return (
+    <Container>
+    <div className="stick_follow_icon">
+      <ul>
+        {socialprofils.twitter && (
+          <li>
+            <a href={socialprofils.twitter}>
+              <FaTwitter />
+            </a>
+          </li>
+        )}
+        {socialprofils.github && (
+          <li>
+            <a href={socialprofils.github}>
+              <FaGithub />
+            </a>
+          </li>
+        )}
+        {socialprofils.facebook && (
+          <li>
+            <a href={socialprofils.facebook}>
+              <FaFacebookF />
+            </a>
+          </li>
+        )}
+        {socialprofils.linkedin && (
+          <li>
+            <a href={socialprofils.linkedin}>
+              <FaLinkedin />
+            </a>
+          </li>
+        )}
+        {socialprofils.youtube && (
+          <li>
+            <a href={socialprofils.youtube}>
+              <FaYoutube />
+            </a>
+          </li>
+        )}
+        {socialprofils.twitch && (
+          <li>
+            <a href={socialprofils.twitch}>
+              <FaTwitch />
+            </a>
+          </li>
+        )}
+      </ul>
+      <p>Follow Me</p>
+    </div>
+    </Container>
+  );
+};
